% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{gbif_dl}
\alias{gbif_dl}
\title{Use rgbif to get a citable GBIF download}
\usage{
gbif_dl(
  taxa,
  kingdom = "all",
  dir.out = "gbif_dl",
  maxrec = Inf,
  gbif_user = "",
  gbif_pw = "",
  gbif_email = ""
)
}
\arguments{
\item{taxa}{A string of the form 'genus species' or 'genus', or a vector of several names in this form.}

\item{kingdom}{Specify ONE taxonomic kingdom for this query. Valid strings are "all", or kingdom names as defined in the GBIF taxonomy.}

\item{dir.out}{Path to your output directory}

\item{maxrec}{Maximum number of records to download.}

\item{gbif_user}{Your GBIF username}

\item{gbif_pw}{Your GBIF password}

\item{gbif_email}{Your GBIF email}
}
\description{
This function requests data from the GBIF database for one taxon or a list of taxa
using the rgbif access to the GBIF API. The raw data are stored and a reformatted data frame
is returned that is compatible with downstream cRacle functions.
}
\details{
Set up a GBIF account for access to downloads before using this function. Then you may either pass
your username (gbif_user), password (gbif_pw), and email (gbif_email) to this function OR write them
to your .Renviron file using:

system('echo "env_gbif_user=\'your_username\'" >> ~/.Renviron');
system('echo "env_gbif_pw=\'your_password\'" >> ~/.Renviron');
system('echo "env_gbif_email=\'your_email@gmail.com\'" >> ~/.Renviron');
system("source~/.Renviron");
}
\examples{
\dontrun{
test <- gbif_dl('Amborella trichopoda');
}
}
\author{
Yaniv Kovalski
}
