% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grab.R
\name{get_dist_all}
\alias{get_dist_all}
\title{Download distribution data from repositories}
\usage{
get_dist_all(taxon, maxrec = 19999, repo = c("gbif"))
}
\arguments{
\item{taxon}{A string of the form 'genus species' or 'genus'.}

\item{maxrec}{Maximum number of records to download.}

\item{repo}{Which data repositor(ies) should be searched. Accepts 'gbif', 'inat', and/or 'bison'. Defaults to 'gbif'.}
}
\description{
This function requests data from the GBIF database for a single taxon using the GBIF callback API.
}
\examples{
\dontrun{
abies <- get_dist_all('Abies', maxrec = 1000);
}
}
