% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probability.R
\name{get_optim}
\alias{get_optim}
\title{Find PDF optim(a)um}
\usage{
get_optim(dens.ob)
}
\arguments{
\item{dens.ob}{An object derived from the cRacle::dens_ob() function.}
}
\value{
A list with letters and numbers.
\itemize{
  \item conintkde - 95% Confidence intervals using kernel density distributions, N-CRACLE
  \item conintgauss- 95% Confidence intervals using gaussian density distribution. P-CRACLE
  \item origk - Kernel density optimal interval as in Harbert & Nixon, 2015. Within 1 percent of the maximal likelihood.
  \item origg - Gaussiandensity optimal interval as in Harbert & Nixon, 2015. Within 1 percent of the maximal likelihood.
  \item dirconint - Direct calculation of the gaussian based confidence interval using weighted means. Should be identical to conintgauss.
  \item means - Weighted means, Should be the same as the maximally likely value of the gaussian density curve.
  \item sds - Weighted standard deviation.
}
}
\description{
Using an object from the cRacle::dens_obj() function. Create a single density object (i.e., like that produced by cRacle::densform()) where the probability curves correspond to the probability density function of ALL taxa/species from the original set occurring.
}
\examples{
\dontrun{
#distr <- read.table('test_mat.txt', head=T, sep ="\t");
#OR:
data(distr);
data(climondbioclim);
extr.raw = extraction(data=distr, clim= climondbioclim, schema='raw');
dens.list.raw <- dens_obj(extr.raw, clim = climondbioclim, bw = 'nrd0', n = 1024);
multiplot(dens.list.raw, names(climondbioclim[[1]]));
and <- and_fun(dens.list.raw);
addplot(and, names(climondbioclim[[1]]), col ='black');
optim.and <- get_optim(and);
abline(v=optim.and$means[paste(names(climondbioclim[[1]]), 'mean', sep = ".") ])
}
}
